requirements.txt
asgiref==3.6.0
decorator==5.1.1
dj-database-url==1.3.0
django
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
httpcore==0.17.0
httpx==0.24.0
sqlparse==0.2.4
uvicorn==0.21.1